Moved Panel to the Main namespace
