+++ src/Hypzeh.Web/ClientApp/components/App/components/layout/index.jsx
@@ -1,11 +1,9 @@
 import Main from './Main';
 import Navbar from './Navbar';
 import Page from './Page';
-import Panel from './Panel';
 
 export {
   Main,
   Navbar,
   Page,
-  Panel,
 };
